+++ lib/NetworkManager.js
@@ -20,12 +20,11 @@ const Multimap = require('./Multimap');
 class NetworkManager extends EventEmitter {
   /**
    * @param {!Puppeteer.CDPSession} client
-   * @param {!Puppeteer.FrameManager} frameManager
    */
-  constructor(client, frameManager) {
+  constructor(client) {
     super();
     this._client = client;
-    this._frameManager = frameManager;
+    this._frameManager = null;
     /** @type {!Map<string, !Request>} */
     this._requestIdToRequest = new Map();
     /** @type {!Map<string, !Protocol.Network.requestWillBeSentPayload>} */
@@ -52,6 +51,13 @@ class NetworkManager extends EventEmitter {
     this._client.on('Network.responseReceived', this._onResponseReceived.bind(this));
     this._client.on('Network.loadingFinished', this._onLoadingFinished.bind(this));
     this._client.on('Network.loadingFailed', this._onLoadingFailed.bind(this));
+  }
+
+  /**
+   * @param {!Puppeteer.FrameManager} frameManager
+   */
+  setFrameManager(frameManager) {
+    this._frameManager = frameManager;
   }
 
   /**
@@ -196,7 +202,7 @@ class NetworkManager extends EventEmitter {
         redirectChain = request._redirectChain;
       }
     }
-    const frame = event.frameId ? this._frameManager.frame(event.frameId) : null;
+    const frame = event.frameId && this._frameManager ? this._frameManager.frame(event.frameId) : null;
     const request = new Request(this._client, frame, interceptionId, this._userRequestInterceptionEnabled, event, redirectChain);
     this._requestIdToRequest.set(event.requestId, request);
     this.emit(NetworkManager.Events.Request, request);
