+++ lib/Page.js
@@ -18,7 +18,6 @@ const fs = require('fs');
 const EventEmitter = require('events');
 const mime = require('mime');
 const {NetworkManager} = require('./NetworkManager');
-const {NavigatorWatcher} = require('./NavigatorWatcher');
 const {Dialog} = require('./Dialog');
 const {EmulationManager} = require('./EmulationManager');
 const {FrameManager} = require('./FrameManager');
@@ -79,16 +78,15 @@ class Page extends EventEmitter {
     this._keyboard = new Keyboard(client);
     this._mouse = new Mouse(client, this._keyboard);
     this._touchscreen = new Touchscreen(client, this._keyboard);
-    /** @type {!FrameManager} */
-    this._frameManager = new FrameManager(client, frameTree, this);
-    this._networkManager = new NetworkManager(client, this._frameManager);
+    this._networkManager = new NetworkManager(client);
+    this._frameManager = new FrameManager(client, frameTree, this, this._networkManager);
+    this._networkManager.setFrameManager(this._frameManager);
     this._emulationManager = new EmulationManager(client);
     this._tracing = new Tracing(client);
     /** @type {!Map<string, Function>} */
     this._pageBindings = new Map();
     this._ignoreHTTPSErrors = ignoreHTTPSErrors;
     this._coverage = new Coverage(client);
-    this._defaultNavigationTimeout = 30000;
     this._javascriptEnabled = true;
     /** @type {?Puppeteer.Viewport} */
     this._viewport = null;
@@ -254,7 +252,7 @@ class Page extends EventEmitter {
    * @param {number} timeout
    */
   setDefaultNavigationTimeout(timeout) {
-    this._defaultNavigationTimeout = timeout;
+    this._frameManager.setDefaultNavigationTimeout(timeout);
   }
 
   /**
@@ -577,42 +575,8 @@ class Page extends EventEmitter {
    * @return {!Promise<?Puppeteer.Response>}
    */
   async goto(url, options = {}) {
-    const referrer = typeof options.referer === 'string' ? options.referer : this._networkManager.extraHTTPHeaders()['referer'];
-
     const mainFrame = this._frameManager.mainFrame();
-    const timeout = typeof options.timeout === 'number' ? options.timeout : this._defaultNavigationTimeout;
-    const watcher = new NavigatorWatcher(this._client, this._frameManager, this._networkManager, mainFrame, timeout, options);
-    let ensureNewDocumentNavigation = false;
-    let error = await Promise.race([
-      navigate(this._client, url, referrer),
-      watcher.timeoutOrTerminationPromise(),
-    ]);
-    if (!error) {
-      error = await Promise.race([
-        watcher.timeoutOrTerminationPromise(),
-        ensureNewDocumentNavigation ? watcher.newDocumentNavigationPromise() : watcher.sameDocumentNavigationPromise(),
-      ]);
-    }
-    watcher.dispose();
-    if (error)
-      throw error;
-    return watcher.navigationResponse();
-
-    /**
-     * @param {!Puppeteer.CDPSession} client
-     * @param {string} url
-     * @param {string} referrer
-     * @return {!Promise<?Error>}
-     */
-    async function navigate(client, url, referrer) {
-      try {
-        const response = await client.send('Page.navigate', {url, referrer});
-        ensureNewDocumentNavigation = !!response.loaderId;
-        return response.errorText ? new Error(`${response.errorText} at ${url}`) : null;
-      } catch (error) {
-        return error;
-      }
-    }
+    return await this._frameManager.navigateFrame(mainFrame, url, options);
   }
 
   /**
@@ -633,17 +597,7 @@ class Page extends EventEmitter {
    */
   async waitForNavigation(options = {}) {
     const mainFrame = this._frameManager.mainFrame();
-    const timeout = typeof options.timeout === 'number' ? options.timeout : this._defaultNavigationTimeout;
-    const watcher = new NavigatorWatcher(this._client, this._frameManager, this._networkManager, mainFrame, timeout, options);
-    const error = await Promise.race([
-      watcher.timeoutOrTerminationPromise(),
-      watcher.sameDocumentNavigationPromise(),
-      watcher.newDocumentNavigationPromise()
-    ]);
-    watcher.dispose();
-    if (error)
-      throw error;
-    return watcher.navigationResponse();
+    return await this._frameManager.waitForFrameNavigation(mainFrame, options);
   }
 
   /**
